optimizer_backend/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.75.0
pydantic==2.11.3
pypdf==5.4.0
python-docx==1.1.2
python-multipart==0.0.20
